for 3. when is a better implementation than a cascaded if according to the solution. IDE did it for me.
for 4. in the words of the solution which said it better than I ever could, "The most succinct solution passes lambda expressions as the arguments, uses the it parameter reference in place of the parameter names, and makes use of trailing lambda syntax." The code by itself removed the need to create a variable as well to create the function fToK after

for 5 create a custom getter that you can then call in main function like so     println(brunoSong.isPopular). 1,000,000 is less preferable than 1_000_000 which is easier to read

while 8 had the correct solution already I enjoyed the explanation so am pasting it here:
The solution uses the ?. safe call operator and the ?: Elvis operator to return the correct price: